refactor: 패키지 구조 변경
계층형 -> 도메인 별로 모아놨다.

컨트롤러, 뷰, 상수를 어디에 두기 애매해서 game 패키지에 넣어뒀는데 조금더 생각해봐야겠다.
