+++ src/main/java/racingcar/generator/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.domain.generator;
+package racingcar.generator;
 
 import camp.nextstep.edu.missionutils.Randoms;
