+++ src/main/java/racingcar/racing/RoundCount.java
@@ -1,6 +1,4 @@
-package racingcar.vo;
-
-import racingcar.exception.RoundCountValidator;
+package racingcar.racing;
 
 public class RoundCount {
